Add files via upload
This is an backgroundScreen generator where the background screen changes automatically according  to user need .This is an project based on Web development using as Html and Css. This I recevied as a  task by #codealpha. this was a great journey to work with them .Felt so Honourable . IDE used is Visual Studio Code.